requirements.txt
black==24.2.0
matplotlib==3.8.3
Pillow==10.2.0
pyntcore==2024.2.1.3
skia_python==121.0b6
streamdeck==0.9.5
